bz18930: Guard against get_total_playback_time() use.
This method may return None, so guard against its usage in arithmetic
operations.  In addition it may theoretically return 0.  Make sure
we never use such a value in the denominator.

Refactored the code in GTK/QuickTime so now everybody uses the same
play_at_time(), skip_forward() and skip_backward() methods.  This involved
backing out part of 0b79397d that reimplemented play_at_time() for GTK, there
is a comment explaining it is for optimization purposes and while I think
the comment is valid I don't think this is a hot path and that we really ought
to worry about it.