bz18806: call get_metadata() lazily for echonest.
If we add a bunch of paths to our echonest processor, then we were calling
get_metadata() on all of them.  If the user clicks the buttons in the
preferences window, this could be thousands of paths.  get_metadata() is too
slow to be called on that many paths at once.  Instead, try to wait to call it
until we really need the data.

This was probably the slowest part of set_net_lookup_enabled()