Made the frontend more robust when handling startup failures.
If we have a startup failure outside of the eventloop stuff (right now this is
only creating the extension manager stuff), then the backend continues to send
messages to the frontend after the StartupFailed message.  The frontend should
just ignore them.

This is really a band-aid for the real problem which is that if the extension
manager breaks, then the backend shouldn't continue doing things.  But I
figure it's better to do it than not.