Changes to the title column of the item table.
Before title tracked the title that we got from metadat (file tags, echonest,
etc).  Now that's stored in metadata_title, and title is the computed value of
title (what get_title() used to return).

This is needed if we are going to implement frontend item list using SQL
queries, since if the user sorts by title, we need this for the ORDER BY
clause.