Make plugins work on linux.
I'm not sure the ramifications of changing gtk.gdk.threads_init() to
gobject.threads_init(), but it seems correct in theory at least.  We don't
need gtk.gdk.threads_init() unless we access GTK/GDK from separate threads,
which we don't.  And if we did that, we would need a
threads_enter()/threads_leave() pair, which we're not using.