Worked on frontend using the items that ItemTracker outputs.
Made an new ItemInfo class that uses data directly from the DB.  I tried to
keep the old API as much as possible, but changed it in a couple of places.

Added a unittest to check that itemrenderer module doesn't throw an error when
using the new ItemInfo.

Reduced the number of items we fetch in one chunk.  After writing this class,
I realized that we need many joins to generate an item, so I think it makes
sense to not try to make so many at one time.