Refactored the logic for newly-available items.
Before we would set last_viewed on our feed, then use that to calculate if the
item was new (plus several other not entirely consistent criteria).  Changed
it so that we simply have a new column in the items table.

This is going to help the frontend data avoid needing to join to the feeds
table just to tell if an item is newly-available.