Unroll remote_downloader.status column
Instead of storing a dictionary using repr, we store each value in its own
column.  This is cleaner in general and will help the frontend data project
since we need to be able to sort on some of these columns.

While we're messing with thing, I tried to refactor some of the logic for
these values.  Mostly I wanted to try to use None consistently to mean an
unknown or invalid value.  Also changed the type of some of the values.

  - eta/rate use None if we can't calculate them
  - endTime is None before the download has finishing, instead of being equal
    to startTime
  - totalSize is None if we don't know it instead of -1
  - retryCount uses None if we aren't retrying downloads instead of -1
  - Made filename/shortFilename initialize to None instead of ""
  - currentSize starts at 0 instead of None
  - downloaders don't send back the channelName value.  They never change it
    so that was just redundent.
  - Made startTime, endTime, eta, and rate integers instead of floats

Changed the hack for torrent magnet link titles.  Instead of setting
item.title, it sets torrent_title.  It also uses the metadata rather than the
filename attribute.  I think this is more correct than before.