Attach a DB to store temporary info.
Before we had a system where we would update the RemoteDownloader objects
in-memory and only save the data every X second or so.  This won't work if the
frontend needs to query that data using sqlite.

Instead, we can create a table with journal_mode=MEMORY and synchronous=OFF.
This should make it very fast, so it should be a reasonable replacement for
the old system.