Ensure that the DB transaction is finished before emitting ItemChanges
Before both LiveStorage and ItemChangeTracker connected to the
eventloop-finished signal.  However, to make things work right we need to
ensure that the ItemChangeTracker code runs after LiveStorage commits its
transaction.  There was no guarentee that that was happening before.

Changed it so that LiveStorage emits a signal when it's done with its
transaction and made ItemChangeTracker connect to that.