Drop the seen column on item.
It's just redundent when we have watched_time.

Dropped references to "seen" and changed them to "watched".  It's what we use
in the frontend and it's more clear that "watched" means the file has been
played, rather then the user has just seen the item in an item list.

Fixed typo that accidentally worked in upgrade182.