Started using the new ItemList in the frontend code.
Lots of refactoring was required.  In particular ItemListController and Item
were changed to work with each other.

Implemented the sorts/filters from the old item list code.  Some of them still
have some work to be done.

Renamed the type of channel folders to "feed-folder".  I think this means that
users will lose their saved settings for those tabs (scroll position, which
list is active, etc).  We could fix this with a DB migration, but I don't
think it's so important.

Removed old TableView code for bulk changes.

Don't add a background widget between StandardView and its container.  I'm not
sure why exactly, but this prevents GTK from using the fixed height
table optimizations.

Made it so we initialize the data package during frontend startup.

Made it so we don't keep around the audio/video tabs forever.  The new code
can fetch the items for those displays fast enough that we don't need this
hack anymore.

Removed old resort_on_update() code.  I'm not sure how it would work with the
new system, and I'm also not sure if it's needed anymore.

Changed ItemList.get_group_info() to return the first item in the group like
the old function used to.