Removing tempdb code.
It was not a great idea to begin with.  Using no journaling mode is
interesting, but we actually prefer WAL mode over it.  Otherwise the
backend can't write to the database until the frontend is done reading it.
This is a performance issue, but it can also cause deadlocks.

We actually still have that same issue on databases where we fail to change to
WAL mode.  Added a FIXME note about this.

Removed the code to delay saving the RemoteDownloader objects to disk.  We
need to update the DB for each status update now.  Removed rate_limit code it
doesn't make sense with the new system.