Simplified feed title handling.
Before it was hard to track where the name of the feed came from.  The code
was spread across Feed.get_title() and the various FeedImpl subclasses
get_title(), then sometimes it was ignored and a blank value was used instead
(for the manual feed).  Hopefully this way is more straightforward.

Made directory feed titles blank, this is consistent with how we handle the
manual feed.

Made to fix for #16044.

Removed the searchtest.  It didn't pass anymore and we're changing the search
code anyways.