Reworked parent_sort_key/feed_name attributes.
The way these worked with the old ItemInfo class was weird.  They both
basically referred to the same thing, even though their names were different.
Also it was pretty complicated to calculate either value.

Changed it so that we store parent_title in the database, which corresponds to
the old feed_name.  Also added the parent_title_for_sort attribute to ItemInfo
which corresponds to the old parent_sort_key.