Database cleanup and improvement
Call PRAGMA wal_checkpoint before copying our database file.  We don't copy
the wal file so we need to ensure that everything has been written from there.

Added public method to LiveStorage to get the sqlite type for a table column.

Added attribute to ConnectionPool that specifies if we are using the WAL
journal mode.

Fixed ConnectionPool.connection_context() method and renamed it to just
context().  It was annoying to type before.