Various fixes/cleanup.
Implemented the others tab.

Added some more info when we get an IndexError in ItemTracker.get_row().

Fixed it so that we don't destroy ItemLists if they are being used by the
playback code.

Refactored ItemFetcherWAL slightly.  It's simpler to just keep the same
connection for as long as the ItemFetcher is alive.