Fix for #19308.
The issue was that we were creating the item list, then immediately calling
set_search(), before we connected to the will-change and list-changed signals.

Changed it so we set the search when we create the ItemList.  Also fixed it so
that we don't call ItemList.set_search() superfluously.  This would happen
when we called self.titlebar.set_search() inside the ItemListController
handler.