Replaced DatabaseItemSource.
I'm planning on removing this code, so the code that used it needed to be
changed.  Added a function to simply fetch ItemInfos by their ids and used
that instead.

I left a few references to DatabaseItemSource around.  That's because it's in
a component that I plan to remove alongside DatabaseItemSource.