Implemented name collation in C.
I noticed that it was fairly slow for me to switch to my music tab when the
items were sorted by name.  Reimplementing it in C made the time go down by
about an order of magnitude.

Made it so we use the binary kits version of sqlite3 on windows.  I just
noticed that the versions from python didn't support FTS3/FTS4.

There are a few differences between this and util.name_sort_key():

- "The" or "A" at the begining of them just get that word choppped off.
  Before, we would append ", The" or ", A" to the end of the name.
- We don't run it through the "NFKD" normalization.  This means we don't
  support accents as well as before.