bz19345: Completely remove delayed sharing start feature.
Don't delay the sharing start in any way until we think about this further.
I think what's happened is we have tried to delete something before the
sharing's started and so it tries to remove an item which the sharing
manager did not know about.

Since sharing wasn't really designed with that in mind let's just disable it
for now since this is the safest.