Don't use the content= sqlite3 parameter
The content option was only added in 3.7.9 (Nov 2011) and officially
documented in 3.7.11 (Mar 2012).  Let's stop using it.  This will result in
larger database files, but I guess we'll have to live with it.

The other option would be to only use it if we had a sqlite version that
supports it, but this is tricky to maintain.  It also makes miro databases
much less portable.