bz19244: Don't create duplicate sqlite databases for a device
This would happen when the frontend sent the backend the SetDeviceType
message.  BackendMessageHandler.handle_set_device_type() would then call
device_connected() on an already connected device and we would create a second
LiveStorage and MetadataManager objects.  Having an extra LiveStoreage object
that we would never close would prevect ejection of the device.