Fix scan_device_for_files()
Changed check from "app.device_manager.running and
os.path.exists(device.mount)" to use _device_not_valid().  We now should quit
if the device is hidden because that means that the sqlite database is closed.

Made it so we return if the check fails.  It seems simpler to have the rest of
the code not have the extra level of indent.