Changed how we handle invalidate_size_request() on OSX
Instead of using a deferred call, just handle it immediately.  I think this is
the simplest way to handle things, the layout process.  We chain
size-request-changed signals to the top-level window which then calls place()
on the content widget for that window and the place() calls cascade down to
the children of each container.

Before there could be issues where we would call invalidate_size_request(),
but defer doing anything.  Before do_invalidate_size_request() was called the
parent of the widget would try to lay it out, which lead to confusion if we
used the wrong size request there.

The new system will occasionally result in nested size-request-changed signals
happenening since place() causes size-allocated and that could trigger
arbitrary code.  I just changed it so that both of those signals were okay to
nest.

Removed the SizeRequestManager code, we don't need it anymore.