bz18863/bz19411: Defer creating a Display object until we need it
Geoff had code that did this before, but it was reverted because it triggered
issues in the OSX layout code.  After the last change it's no longer a
problem.  Re-implementing the solution in a slightly different way than Geoff
did.  This one only uses timer.add() rather than both timer.add() and
call_on_ui_thread().

This also solves 19411.  The issue there was that we were could create many
Display objects, but defer changing to them.  Since the life of an ItemTracker
is tied to the life of a Display, this would result in lots of ItemTrackers
around when we should only have 2 or 3 at most.