Added code to upgrade device databases from version 5.x.
This makes things a bit more complex because we need to worry about upgrading
from both 4.x versions and 5.x versions.  Changed the devicedatabaseupgrade
code from a function object to a regular class.  It needed to be changed to
deal with the extra complexity.

Made it so that database upgrade funcs conditionally run based on if the
database is the main DB or a device database.

Added unittests to check things.

Changed the linux setup.py to copy the entire testdata tree to the resources
directory.  This way we don't need to keep modifying it every time we add a
subdirectory.