Implemented DeviceItemTracker.
Wrote DeviceItemInfo, which subclasess ItemInfo, but is created from a
different SQL query.

We now create a ConnectionPool object for each connected device on the
frontend.  There should only be a connection while the tab is being displayed,
so it shouldn't interfere with ejection.

We also create a subclass of ItemChangeTracker that handles changes for all
device items.

Fixed a couple bugs with the ConnectionPool._check_wal_mode() logic.