Refactored item/itemtrack to make them more flexible.
This is all preparation for making a version of ItemTracker that works with
device databases.  For those, we are using the device_item table instead of
the item table and also we aren't joining to the tables like
remote_downloader and icon_cache.

The steps to make ItemTracker work with a device database are:

- Subclass ItemSelectInfo and define the select for device databases
- Subclass ItemInfo and override the select_info attribute
- Subclass ItemTracker and override the item_info_class attribute