Refactored the ItemTracker class hierarchy.
We were going to run into problems because the frontend subclasses
ItemTracker, but we were also subclassing it for DeviceItemTracker.  We could
solve that with a mixin classes, but that doesn't seem very clean to me.

Instead, made it so ItemTrackerQuery is the object that stores the which
ItemInfo/ItemSelectInfo classes we are going to use.  This makes sense, since
the ItemTrackerQuery is needs that info.

The result is that we can subclass ItemTracker with less pain in the frontend.

Moved the code that handles keeping a connection pool for each device into
connectionpool.py.  That doesn't need to be inside the ItemTracker class.