Started showing item lists for devices.
Added DeviceItemList class and used that to handle the data.

Created DeviceLiveStorage class that modifies LiveStorage slightly to not
create a full text search table.  In the future we should change this to make
one that uses device_item as its content source.

Fixed a bunch of bugs that showed up when I tried to use the device code with
a real database.