Made the ItemSource code work with the new item infos
Created version of ItemSource that works with the new ItemInfo system.  It
works like the old ItemSource class in that it tries to abstract away the
differences between regular databases and device databases.

Added an item_source attribute to ItemInfo and device_id to DeviceItemInfo.
Fixed DeviceItem so that we set device_id when we restore items.

Fixed ItemInfo.__str__ method.