Fixed DeviceConnectionPoolMap to handle changed DeviceInfos
This handles the case where miro starts up and the device isn't mounted, then
the user mounts it.  In that case DeviceConnectionPoolMap first sees the
DeviceInfo when it's part of the added list, but at that point db_info=None.
When it sees it's part of the changed list and db_info is no longer None, then
it should create a connection pool.