bz19484: fix setting small width mode.
The issue was that when we would call set_small_width_mode() that would
trigger a bunch of size-request-changed signals and then we would get the
size-allocated signal and call set_small_width_mode() again.  The fix is to
set in_small_width_mode before changing the widgets so when we see the
size-allocated signal we don't do anything.