Made SharingItemTrackerImpl create an sqlite database
The trickiest thing was to decide how to handle the acutal files for
sharing DBs, since they should be just temporary.  The situation I
wanted to avoid was miro crashing and leaving behind temporary files.
Decided to go with a scheme where use the path
%support_dir%/sharing-db-X and we try to delete exiting
files if we're not currently using them for a sharing DB.

Other than that, nothing is really implemented.  We just open the
database when we connect to a share and close it at disconnection.

Added unittests to check.  Refactored SharingItemTrackerImpl slightly
to make it easier to test.