bz19039: check that we create widgets inside the UI thread
Moved code that checks if we are running in the event loop thread from
database.py to a new module.  Added code to check if we are running in the UI
thread as well.

This code has the side effect of raising an exception if a widget gets created
before the gtk.main() or NSApplicationMain() is called.  I'm pretty sure we've
had issues in the past when we didn't do this.