Created backend class to track data for a share
This solves a bunch of issues with getting the new code to work:
  - We need to send the frontend the path to the database, but it
    didn't exist when SharingInfo was created.
  - We weren't calling SharingTracker.get_tracker() anymore, so
    SharingItemsTrackerImpl was not going to be created.

I think it's also a step in the right direction to create a
backend-only Share object.  Right now we have a bad situation where
both the backend and the frontend are using the same SharingInfo
object.  There's some synchronization issues there and also, I think
some things should be only accessible from the backend.  Ticket #19869
has more of my thoughts on this.

Added sqlite_path to SharingInfo.  The frontend is going to need it
once it starts using the itemtrack code.

Changed the code to create databases slightly:
  - Don't keep trying to create databases if opening the database
    fails.  If we can't open a sqlite database on an empty path then
    we probably shouldn't just keep trying new paths.  It's probably a
    permissions error, or a disk space issue.
  - Use fileutil.delete() to delete the file.  This works around
    windows file locking issues and if the frontend still has the
    database open.
  - Because we are now not sure when a database file will get deleted,
    don't try to re-use paths that we've already used this run.