Implemented the frontend-side for the new sharing items system
Made SharingItem.video_path a unicode value rather than a string one.
This matches the old API, so I guess it's best.

Changed how we compute sharing ids.  Before they were tuples, which is
a bit weird when used with the rest of the frontend code.  We now
convert the tuple to a int by using hash().  Also changed how we store
ids for SharingInfo/PlaylistInfo.  The new ids are
"sharing-<share_id>" and "sharing-<share_id>-<playlist_id>"
respectively.

Made it so that DeviceItemInfo and SharingItemInfo don't inherit
directly from ItemInfo.  This avoids them picking up properties that
were created based on the ItemInfo's select_info.

Changed SharingItemChanges slightly.  Now playlist_changed is just a
bool that stores if any playlists have changed.  I think that's how we
were using it in the frontend anyway.

Fixed the ConnectionPoolTracker code.  Before it was assuming we got a
info class when things get removed, but we just get the tab id.
Changed things so that we just use the tab id as the key.  I'm pretty
sure tab ids have to be unique anyways.

Cleaned up share ejection.  Switched to using the StopTrackingShare
message instead of SharingEject.  They basically do the same thing and
the new message works better with the rest of the code.  Clear out the
database in stop_tracking().

The one thing not implemented yet is the playlist/podcast tabs under
the share.