Added BackendItemTracker
This will be used to track items for the sharing server.

Added __hash__/__eq__ methods to ItemInfo

Allow ItemTrackerQuery objects to not have an order by clause

Added ItemTrackerQuery method to select all the row data, not just the
ids.

Made ItemTrackerQuery use LEFT JOIN instead fo JOIN to join to tables
that we use for conditions.  This add flexibility, for example
conditions that use OR probably want a left join.

Made LiveStorage.execute a public method.