Rewrote the sharing server to use the new ItemTracker system
Changed the sharing test to implement what I think is the correct
behavior, which is that get_items() with no playlist will return items
that are in playlists, even if SHARE_AUDIO/SHARE_VIDEO is not set.

Fixed ItemTracker so that it tracks when the join column changes for
conditions in tabled joined to the item table.