Refactored the infoupdater since we no longer use it for items.
The last bit of code that used it was in frontend/gtk/video, but we
didn't need it anymore since connecting to the "playing-info-changed"
signal works fine.

Fixed PlaybackPlaylist code so that it updates currently_playing when
items change.

Removed code that creates app.info_updater in the cli/shell
applications.  That was needed for the sharing code, but not anymore.
Moved the module back into the widgets package.