Database Error refactoring
Added code to handle database errors in the frontend as well as the
backend.  This is pretty tricky for a couple reasons:

- Errors can happen in both threads at the same time, or near in time
  to one and other.  In that case, we don't want to show 2 dialog
  windows.
- On GTK at least, we can have additional errors happen while waiting
  for the user choice for our first dialog window (for example, a
  timeout fires for the ItemTracker code).

Here's how we handle it:
- The backend system stays pretty much the same.  It asks the frontend
  to show a dialog and blocks until that dialog is responded to.  The
  only difference is that we use a special Dialog subclass.
- If we get an error in the frontend, then we show a similar dialog
  and if the user selects retry we run a callback function to recover
  from the error.  This should handles nested errors okay.  (The
  actual recovery code is yet to be written.)
- We avoid showing multiple dialog windows at once and instead reuse the
  response from the first dialog.

Added a frontend class that defines a generic API that frontends can
present to the portable code.  Right now it's just one function, but
it seems like it could be useful now that we have more portable code
running in the frontend thread.

Slight fix for show_choice_dialog().  If the user closes the dialog,
then we return None.

Made the backend code always check for sqlite3.DatabaseError.  I think
the best error since it's very general, but won't catch interface
errors (for example syntax errors).