Refactored the startup checks code
I think it makes sense to put them in the same class since they're so
closely tied together.  Also made it so in the normal case, we don't
use any idle callbacks.  I'm not sure if this will make a big
difference, but it could be faster and it doesn't hurt anything.