Startup cleanup phase one.
Moved a bunch of code to run later on in the startup process.  The
idea is to make as little as possible run in the backend until we have
our window up.

Renamed some of the functions in startup.py and added docstrings.

Don't setup app.transcode_manager.  We don't actually use that
anywhere.