bz19904: fix titles for youtube videos.
Don't reset torrent_title to None for HTTP downloads, we use that for
youtube videos to set the title of the item.

Changed it so that if torrent_title gets changed, then we call
calc_title() automatically.  Before we weren't doing that so the title
stuck around until the download completed.  This fixes #19900